Move interfaces and change namespaces
