+++ src/Social/Contract/Interactions/InteractionTypesInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Kanvas\Packages\Social\Contract\Interfaces;
+namespace Kanvas\Packages\Social\Contract\Interactions;
 
 interface InteractionTypesInterface
 {
