+++ src/Social/Contract/Users/UserInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Kanvas\Packages\Social\Contract\Interfaces;
+namespace Kanvas\Packages\Social\Contract\Users;
 
 interface UserInterface
 {
